Move ProviderNodes to InfraProvider
